// Projects/frontend/src/types/user.ts
export interface User {
  id: number;
  email: string;
  name: string;
}

export interface LoginDTO {
  email: string;
  password: string;
}

export interface RegisterDTO {
  email: string;
  password: string;
  name: string;
}

export interface AuthResponse {
  access_token: string;
  user: User;
}